v2.00rc1

+++ WEB-INF/classes/gov/noaa/pfel/erddap/dataset/GridDataRandomAccessor.java
@@ -47,8 +47,6 @@ public class GridDataRandomAccessor {
      */
     public static boolean verbose = false; 
 
-    //things passed into the constructor
-    protected GridDataAccessor gridDataAccessor;
     
     //things the constructor generates
     protected NDimensionalIndex gdaTotalIndex;
@@ -63,34 +61,35 @@ public class GridDataRandomAccessor {
      * @param tGridDataAccessor a rowMajor gridDataAccessor
      * @throws Throwable if trouble
      */
-    public GridDataRandomAccessor(GridDataAccessor tGridDataAccessor) throws Throwable {
+    public GridDataRandomAccessor(GridDataAccessor gridDataAccessor) throws Throwable {
+        try {
+            if (!gridDataAccessor.rowMajor())
+                throw new Exception("GridDataRandomAccessor.constructor requires the gridDataAccessor to be rowMajor.");
 
-        gridDataAccessor = tGridDataAccessor;
-        if (!gridDataAccessor.rowMajor())
-            throw new Exception("GridDataRandomAccessor.constructor requires the gridDataAccessor to be rowMajor.");
+            //make the dataRaf's
+            EDV dataVars[] = gridDataAccessor.dataVariables();
+            int nDv = dataVars.length;
+            dataRaf = new RandomAccessFile[nDv];
+            dataClass = new Class[nDv];
+            String tQuery = gridDataAccessor.userDapQuery();
+            rafName = gridDataAccessor.eddGrid().cacheDirectory() + //dir created by EDD.ensureValid
+                String2.md5Hex12(tQuery == null? "" : tQuery) + "_" +
+                Math2.random(Integer.MAX_VALUE) + "_";
+            for (int dv = 0; dv < nDv; dv++) {
+                dataClass[dv] = dataVars[dv].destinationDataTypeClass();
+                dataRaf[dv] = new RandomAccessFile(rafName + dv, "rw");            
+            }
 
-        //make the dataRaf's
-        EDV dataVars[] = gridDataAccessor.dataVariables();
-        int nDv = dataVars.length;
-        dataRaf = new RandomAccessFile[nDv];
-        dataClass = new Class[nDv];
-        String tQuery = gridDataAccessor.userDapQuery();
-        rafName = gridDataAccessor.eddGrid().cacheDirectory() + //dir created by EDD.ensureValid
-            String2.md5Hex12(tQuery == null? "" : tQuery) + "_" +
-            Math2.random(Integer.MAX_VALUE) + "_";
-        for (int dv = 0; dv < nDv; dv++) {
-            dataClass[dv] = dataVars[dv].destinationDataTypeClass();
-            dataRaf[dv] = new RandomAccessFile(rafName + dv, "rw");            
+            //get all the data
+            while (gridDataAccessor.increment()) {
+                for (int dv = 0; dv < nDv; dv++)
+                   PrimitiveArray.rafWriteDouble( //this doesn't work for Strings
+                       dataRaf[dv], dataClass[dv], gridDataAccessor.getDataValueAsDouble(dv));
+            }
+            gdaTotalIndex = gridDataAccessor.totalIndex();
+        } finally {
+            gridDataAccessor.releaseGetResources();
         }
-
-        //get all the data
-        while (gridDataAccessor.increment()) {
-            for (int dv = 0; dv < nDv; dv++)
-               PrimitiveArray.rafWriteDouble( //this doesn't work for Strings
-                   dataRaf[dv], dataClass[dv], gridDataAccessor.getDataValueAsDouble(dv));
-        }
-        gdaTotalIndex = gridDataAccessor.totalIndex();
-
     }
 
 
@@ -123,12 +122,16 @@ public class GridDataRandomAccessor {
         //and reuse the data (File2.touch() the files) 
         //if another identical request arrives before files are deleted.
         try {
-            int nDv = dataRaf.length;
-            for (int dv = 0; dv < nDv; dv++) {
-                try {
-                    dataRaf[dv].close(); //will already by closed if this method already called           
-                } catch (Throwable t2) {
-                    //String2.log(MustBe.throwableToString(t2));
+            if (dataRaf != null) {
+                int nDv = dataRaf.length;
+                for (int dv = 0; dv < nDv; dv++) {
+                    try {
+                        if (dataRaf[dv] != null)
+                            dataRaf[dv].close(); //will already by closed if this method already called           
+                        dataRaf[dv] = null;
+                    } catch (Throwable t2) {
+                        //String2.log(MustBe.throwableToString(t2));
+                    }
                 }
             }
         } catch (Throwable t) {
@@ -136,37 +139,37 @@ public class GridDataRandomAccessor {
     }
 
     /** 
-     * This closes and deletes the files.
+     * This closes and deletes all resources.
      * It is recommended, but not required, that users of this class call this 
      * (or close) when they are done using this instance.
      * This won't throw an Exception.
      * 
      */
-    public void closeAndDelete() {
+    public void releaseResources() {
         close();
         try {
-            int nDv = dataRaf.length;
-            for (int dv = 0; dv < nDv; dv++) {
-                try {
-                    File2.delete(rafName + dv);
-                } catch (Throwable t2) {
-                    //String2.log(MustBe.throwableToString(t2));
+            if (dataRaf != null) {
+                int nDv = dataRaf.length;
+                for (int dv = 0; dv < nDv; dv++) {
+                    try {
+                        File2.delete(rafName + dv);
+                    } catch (Throwable t2) {
+                        //String2.log(MustBe.throwableToString(t2));
+                    }
                 }
+                dataRaf = null;
             }
         } catch (Throwable t) {
         }
     }
 
     /** 
-     * Users of this class shouldn't call this -- use close() instead.
+     * Users of this class shouldn't call this -- use releaseResources() instead.
      * Java calls this when an object is no longer used, just before garbage collection. 
      * 
      */
     protected void finalize() throws Throwable {
-        try {  //extra insurance
-            close();
-        } catch (Throwable t) {
-        }
+        releaseResources();
         super.finalize();
     }
 }
